Restructured classes into separate files
Complete restructure of source code. No code changes.
