+++ core/core.h
@@ -1,0 +1,10 @@
+#ifndef CPPDATALIB_CORE_CORE_H
+#define CPPDATALIB_CORE_CORE_H
+
+#include "base64.h"
+#include "hex.h"
+#include "fp_convert.h"
+// value_builder.h includes stream_base.h and value.h
+#include "value_builder.h"
+
+#endif // CPPDATALIB_CORE_CORE_H
