+++ cppdatalib.h
@@ -1,0 +1,16 @@
+#ifndef CPPDATALIB_H
+#define CPPDATALIB_H
+
+#include "core/core.h"
+#include "json/json_pointer.h"
+#include "json/json.h"
+#include "bencode/bencode.h"
+#include "property_list/plain_text.h"
+#include "property_list/xml.h"
+#include "rpc/xml.h"
+#include "csv/csv.h"
+#include "binn/binn.h"
+#include "ubjson/ubjson.h"
+#include "netstrings/netstrings.h"
+
+#endif // CPPDATALIB_H
